% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickGroupMethod.R
\name{pickGroupMethod}
\alias{pickGroupMethod}
\title{pickGroupMethod}
\usage{
pickGroupMethod(data, continList = NULL, catList = NULL, grouping = NULL,
  minCount = NULL, minExpectedCountProportion = 0.75,
  maxLevelsCutoff = NULL, dropOneRow = T, dropNLessThan = 1,
  testNormality = T, minDiscreteValues = NULL, varEqual = T,
  verbose = T)
}
\arguments{
\item{data}{data.frame containing all data}

\item{continList}{continList Vector of continuous variable names}

\item{catList}{catList Vector of categorical variable names}

\item{grouping}{grouping Grouping variable to test for equal variance and other multiple group based assumptions}

\item{minCount}{Minimum total count}

\item{minExpectedCountProportion}{Minimum proportion of cells that have an expected count of 5 or more to still use Chi.square test. Default of 75%}

\item{maxLevelsCutoff}{Maximum number of unique levels to a variable to still analyze. If the variable exceeds this cutoff it will be dropped from the analysis lists.}

\item{dropOneRow}{If TRUE, any variables with only 1 unique level will be dropped from analysis lists.}

\item{dropNLessThan}{Minimum number non-NA responses a variable must have before being dropped from the analysis lists. Default of 1}

\item{testNormality}{If TRUE, normality assumptions are checked for continuous variables}

\item{minDiscreteValues}{minDiscreteValues Minimum number of discrete values of a continuous variable to still be tested with parametric methods.}

\item{varEqual}{If TRUE, and grouping is non-NULL, equal variance assumptions will be tested}

\item{verbose}{If TRUE, messages about decision criteria will be printed to the console.}
}
\value{
List of vectors indicating whether variances should be tested parametrically, non-parametrically or dropped, for continuous and categorical lists separately
}
\description{
pickGroupMethod will test common assumptions for parametric and non-parametric methods for 2 + group methods (e.g. t.test, mann-whitney U, chisq, fisher, ANOVA) and will decide which method to use for each variable.
}
\examples{
#NULL
}

