% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBadVars.R
\name{removeBadVars}
\alias{removeBadVars}
\title{removeBadVars}
\usage{
removeBadVars(data, catlist, predlist, nLevelsCutoff = 5, minCount = 1)
}
\arguments{
\item{data}{input data.frame}

\item{catlist}{vector of categorical variable names to use in multLogistic()}

\item{predlist}{vector of all variable names to use in multLogistic()}

\item{nLevelsCutoff}{Maximum number of levels a categorical variable can have and still be included in the analysis.}

\item{minCount}{Minimum total number of non-NA entries a variable can have and still be included in the analyses.}
}
\value{
a list of varibale name vectors to include in a later analysis.
}
\description{
removeBadVars checks the variables in a dataframe, and removes those that do not meet the given parameters. To be used in conjunction with the multLogistic() function.
}
\examples{
#NULL
}

